make tun as optional dependancy
