+++ easytier/Cargo.toml
@@ -84,7 +84,9 @@ http = { version = "1", default-features = false, features = [
 tokio-rustls = { version = "0.26", default-features = false, optional = true }
 
 # for tap device
-tun = { package = "tun-easytier", version = "0.6.1", features = ["async"] }
+tun = { package = "tun-easytier", version = "0.6.1", features = [
+    "async",
+], optional = true }
 # for net ns
 nix = { version = "0.27", features = ["sched", "socket", "ioctl"] }
 
@@ -188,13 +190,23 @@ defguard_wireguard_rs = "0.4.2"
 
 
 [features]
-default = ["wireguard", "mimalloc", "websocket", "smoltcp"]
-full = ["quic", "websocket", "wireguard", "mimalloc", "aes-gcm", "smoltcp"]
+default = ["wireguard", "mimalloc", "websocket", "smoltcp", "tun"]
+full = [
+    "quic",
+    "websocket",
+    "wireguard",
+    "mimalloc",
+    "aes-gcm",
+    "smoltcp",
+    "tun",
+]
 mips = ["aes-gcm", "mimalloc", "wireguard"]
+bsd = ["aes-gcm", "mimalloc", "smoltcp"]
 wireguard = ["dep:boringtun", "dep:ring"]
 quic = ["dep:quinn", "dep:rustls", "dep:rcgen"]
 mimalloc = ["dep:mimalloc-rust"]
 aes-gcm = ["dep:aes-gcm"]
+tun = ["dep:tun"]
 websocket = [
     "dep:tokio-websockets",
     "dep:http",
