+++ easytier/src/instance/mod.rs
@@ -1,4 +1,7 @@
 pub mod instance;
 pub mod listeners;
+
+#[cfg(feature = "tun")]
 pub mod tun_codec;
+#[cfg(feature = "tun")]
 pub mod virtual_nic;
